back/ceara_api/pyproject.toml
[project]
name = "ceara-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Davi Moreira",email = "davimoreirax2021@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "ceara_api", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.7"
pylint-django = "^2.6.1"
taskipy = "^1.14.1"
pre-commit = "^4.2.0"


[tool.taskipy.tasks]
qos = { cmd = "black src & isort src & pylint src", help = "Executa todas as ferramentas de qualidade de código."}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
